* Only cleanup after tests if they succeeded  * Add automagic database dropping to after-tests cleanup
git-svn-id: svn+ssh://svn.bestpractical.com/svn/jifty.org/jifty/trunk@410 e84bef0a-9b06-0410-84ba-c4c9edb13aeb

+++ lib/Jifty/Test.pm
@@ -159,7 +159,23 @@ sub messages {
 } 
 
 END {
-    unlink mailbox();
+    my $Test = Jifty::Test->builder;
+    # Such a hack -- try to detect if this is a forked copy and don't
+    # do cleanup in that case.
+    return if $Test->{Original_Pid} != $$;
+
+    # If all tests passed..
+    unless (grep {not $_} $Test->summary) {
+        # Clean up mailbox
+        unlink mailbox();
+
+        # Remove testing db
+        Log::Log4perl->get_logger("SchemaTool")->less_logging(3);
+        my $schema = Jifty::Script::Schema->new;
+        $schema->{drop_database} = 1;
+        $schema->run;
+        Log::Log4perl->get_logger("SchemaTool")->more_logging(3);
+    }
 }
 
 1;
